Move retry token logic into a dedicated module
